コンポーネント分割２
頒布物管理画面，レジ画面のコンポーネント分割を実装しました．これで現在実装している全画面の分割が終了し，管理が容易になったと思います
